@@ -72,7 +72,7 @@
 		"pear/pear-core-minimal": "1.10.16",
 		"pear/pear_exception": "1.0.2",
 		"pimple/pimple": "3.5.0",
-		"psr/container": "1.1.2",
+		"psr/container": "2.0.2",
 		"psr/http-client": "1.0.3",
 		"psr/http-factory": "1.1.0",
 		"psr/http-message": "1.1",
@@ -95,7 +95,7 @@
 		"symfony/polyfill-php82": "1.32.0",
 		"symfony/polyfill-php83": "1.32.0",
 		"symfony/process": "5.4.47",
-		"symfony/service-contracts": "2.4.1",
+		"symfony/service-contracts": "3.6.0",
 		"symfony/string": "5.4.47",
 		"symfony/translation": "5.4.45",
 		"symfony/translation-contracts": "2.5.4",
