@@ -15,8 +15,6 @@ use Psr\Container\ContainerInterface;
 use Symfony\Contracts\Service\Attribute\Required;
 use Symfony\Contracts\Service\Attribute\SubscribedService;
 
-trigger_deprecation('symfony/contracts', 'v3.5', '"%s" is deprecated, use "ServiceMethodsSubscriberTrait" instead.', ServiceSubscriberTrait::class);
-
 /**
  * Implementation of ServiceSubscriberInterface that determines subscribed services
  * from methods that have the #[SubscribedService] attribute.
@@ -24,14 +22,12 @@ trigger_deprecation('symfony/contracts', 'v3.5', '"%s" is deprecated, use "Servi
  * Service ids are available as "ClassName::methodName" so that the implementation
  * of subscriber methods can be just `return $this->container->get(__METHOD__);`.
  *
- * @property ContainerInterface $container
- *
  * @author Kevin Bond <kevinbond@gmail.com>
- *
- * @deprecated since symfony/contracts v3.5, use ServiceMethodsSubscriberTrait instead
  */
-trait ServiceSubscriberTrait
+trait ServiceMethodsSubscriberTrait
 {
+    protected ContainerInterface $container;
+
     public static function getSubscribedServices(): array
     {
         $services = method_exists(get_parent_class(self::class) ?: '', __FUNCTION__) ? parent::getSubscribedServices() : [];
